fix camera smooth config issue

+++ Assets/Scripts/Configs/NoneVRSettingManager.cs
@@ -38,7 +38,7 @@ public class NoneVRSettingManager : MonoBehaviour
         if (JsonConfig.HasKey("NVRMode"))
             Dropdown.value = JsonConfig.GetInt("NVRMode");
         if (JsonConfig.HasKey("NVRCameraSmooth"))
-            CameraSmooth.smoothSpeed = JsonConfig.GetInt("NVRCameraSmooth");
+            CameraSmooth.smoothSpeed = (float)JsonConfig.GetDouble("NVRCameraSmooth");
         SetNVRMode();
     }
     public void GetNVRFOV()
@@ -88,7 +88,7 @@ public class NoneVRSettingManager : MonoBehaviour
                 break;
         }
         JsonConfig.SetInt("NVRMode", Dropdown.value);
-        JsonConfig.SetInt("NVRCameraSmooth", Dropdown.value);
+        JsonConfig.SetDouble("NVRCameraSmooth", Dropdown.value);
     }
     public void SetNVRFOV(float fov)
     {
